Moved SettingsDataStorage file to Settings folder
